feat: BROKEN WIP
migrate home page to app dir
