+++ src/core/layouts/Layout.tsx
@@ -1,22 +1,37 @@
-import Head from "next/head"
-import React from "react"
 import { BlitzLayout } from "@blitzjs/next"
-import { useRouter } from "next/router"
+import Head from "next/head"
+import { usePathname } from "next/navigation"
+import React from "react"
 
 export type LayoutProps = { title: string; children?: React.ReactNode }
 
 const Layout: BlitzLayout<LayoutProps> = ({ title, children }) => {
-  const router = useRouter()
-  const canonical = `https://ladderly.io${router.asPath}`
+  const pathname = usePathname()
+  const canonical = `https://ladderly.io${pathname}`
+
   const preppendedTitle = `ladderly.io | ${title}`
 
   return (
-    <>
+    <html lang="en">
       <Head>
         <title>{preppendedTitle}</title>
-        <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
-        <link rel="icon" type="image/png" sizes="32x32" href="/favicon-32x32.png" />
-        <link rel="icon" type="image/png" sizes="16x16" href="/favicon-16x16.png" />
+        <link
+          rel="apple-touch-icon"
+          sizes="180x180"
+          href="/apple-touch-icon.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="32x32"
+          href="/favicon-32x32.png"
+        />
+        <link
+          rel="icon"
+          type="image/png"
+          sizes="16x16"
+          href="/favicon-16x16.png"
+        />
         <link rel="manifest" href="/site.webmanifest" />
         <link rel="canonical" href={canonical} key="canonical" />
 
@@ -25,6 +40,7 @@ const Layout: BlitzLayout<LayoutProps> = ({ title, children }) => {
 
         <meta property="og:type" content="website" />
         <meta property="og:title" content={preppendedTitle} />
+        <meta property="og:url" content="https://ladderly.io/" />
 
         <meta
           name="description"
@@ -34,11 +50,9 @@ const Layout: BlitzLayout<LayoutProps> = ({ title, children }) => {
           name="keywords"
           content="Ladderly, learning, innovation, tech career, open source, community, education, programming, career development, diversity, personalized learning, high-performance lifestyle, tech tools, educational programs, coding, fullstack, full stack, full-stack, javascript, typescript, coding, software engineer, web development, react, nextjs, career, job, remote work, mentorship, mentor, tutorial"
         />
-        <meta property="og:url" content="https://ladderly.io/" />
       </Head>
-
-      {children}
-    </>
+      <body>{children}</body>
+    </html>
   )
 }
